# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_ayMissouri_watchlist-go_git_internal_models.BulkDeleteRequest:
    properties:
      ids:
        items:
          type: string
        type: array
    type: object
...
  github_com_ayMissouri_watchlist-go_git_internal_models.PaginationMeta:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  github_com_ayMissouri_watchlist-go_git_internal_models.PlaysResponse:
    properties:
      plays:
        type: integer
      status:
        $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.WatchlistStatus'
    type: object
...
  github_com_ayMissouri_watchlist-go_git_internal_models.UpdateStatusRequest:
    properties:
      status:
        $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.WatchlistStatus'
    type: object
...
  github_com_ayMissouri_watchlist-go_git_internal_models.WatchlistResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.WatchlistItem'
        type: array
      pagination:
        $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.PaginationMeta'
    type: object
...
host: localhost:8080
info:
  contact:
    name: ayMissouri
    url: https://github.com/ayMissouri/watchlist-go.git
  description: Backend API for tracking movies and TV shows.
  license:
    name: MIT
  title: Watchlist API
  version: "1.0"
paths:
  /anime/watchlist:
    delete:
      consumes:
      - application/json
      description: Deletes several items in one go and tells you how many actually
        went.
      parameters:
      - description: List of item IDs to delete
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.BulkDeleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              format: int64
              type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Bulk delete watchlist items
      tags:
      - watchlist
    get:
      description: |-
        The user's watchlist, one page at a time. Filter by type or status, sort however you like.
        Anime has its own watchlist under /anime/watchlist, with the same endpoints.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: per_page
        type: integer
      - description: Filter by type (not used by the anime watchlist)
        enum:
        - tv
        - movie
        in: query
        name: type
        type: string
      - description: Filter by status
        enum:
        - watching
        - watched
        - plan_to_watch
        - paused
        - dropped
        in: query
        name: status
        type: string
      - description: Sort field
        enum:
        - last_updated
        - title
        in: query
        name: sort
        type: string
      - description: Sort order
        enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.WatchlistResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get watchlist
      tags:
      - watchlist
  /anime/watchlist/{id}:
    delete:
      description: Deletes a watchlist item by ID
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Remove watchlist item
      tags:
      - watchlist
    get:
      description: Returns a single watchlist item by ID
      parameters:
      - description: Item ID (e.g. t63174 or m533535)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.WatchlistItem'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get watchlist item
      tags:
      - watchlist
    put:
      consumes:
      - application/json
      description: |-
        Adds the item, or replaces the whole thing if it's already there. Use the progress/status endpoints for partial updates.
        The anime watchlist takes type "anime", with the MyAnimeList id in `mal_id`. Item ids are shared
        by both watchlists, so an id already used in the other one is rejected with a 409.
      parameters:
      - description: Item ID (e.g. t63174 or m533535)
        in: path
        name: id
        required: true
        type: string
      - description: Watchlist item
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.WatchlistItem'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add or update watchlist item
      tags:
      - watchlist
  /anime/watchlist/{id}/plays:
    delete:
      description: |-
        Drops the most recent play and the watch time that came with it. The item goes
        back to plan_to_watch once no plays are left.
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.PlaysResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Undo the newest play
      tags:
      - watchlist
    post:
      description: |-
        Records one more completed watch of an item and marks it watched. This is the
        "watch it again" half of the checkbox prompt.
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.PlaysResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Log another play
      tags:
      - watchlist
  /anime/watchlist/{id}/progress:
    patch:
      consumes:
      - application/json
      description: Just bumps playback progress. Everything else on the item stays
        as it was.
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      - description: Progress update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.UpdateProgressRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update item progress
      tags:
      - watchlist
  /anime/watchlist/{id}/status:
    patch:
      consumes:
      - application/json
      description: Sets the watch status of an item (watching, watched, plan_to_watch,
        paused, dropped)
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      - description: Status update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.UpdateStatusRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update item status
      tags:
      - watchlist
...
  /watchlist:
    delete:
      consumes:
      - application/json
      description: Deletes several items in one go and tells you how many actually
        went.
      parameters:
      - description: List of item IDs to delete
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.BulkDeleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              format: int64
              type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Bulk delete watchlist items
      tags:
      - watchlist
    get:
      description: |-
        The user's watchlist, one page at a time. Filter by type or status, sort however you like.
        Anime has its own watchlist under /anime/watchlist, with the same endpoints.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: per_page
        type: integer
      - description: Filter by type (not used by the anime watchlist)
        enum:
        - tv
        - movie
        in: query
        name: type
        type: string
      - description: Filter by status
        enum:
        - watching
        - watched
        - plan_to_watch
        - paused
        - dropped
        in: query
        name: status
        type: string
      - description: Sort field
        enum:
        - last_updated
        - title
        in: query
        name: sort
        type: string
      - description: Sort order
        enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.WatchlistResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get watchlist
      tags:
      - watchlist
...
  /watchlist/{id}/progress:
    patch:
      consumes:
      - application/json
      description: Just bumps playback progress. Everything else on the item stays
        as it was.
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      - description: Progress update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.UpdateProgressRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update item progress
      tags:
      - watchlist
...
securityDefinitions:
  BearerAuth:
    description: 'Enter your JWT token as: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"